requirements.txt
faiss-cpu
numpy==1.26.3
openai==1.66.2
scikit-learn
sentence-transformers==3.3.1
tenacity==8.2.3
tiktoken==0.5.1 
torch
#transformers==4.38.1
umap-learn==0.5.5
urllib3==1.26.6 
nltk==3.9.1
psutil==6.1.1
beautifulsoup4==4.13.3
pandas==2.2.3
evaluate==0.4.3
absl-py==2.2.1
rouge_score==0.1.2